Move manifest packageConfig mocks into security_solution plugin (#72527)
